Implements core validations

+++ common/src/main/java/otp/mobile/backend/common/domain/Event.java
@@ -2,17 +2,17 @@ package otp.mobile.backend.common.domain;
 
 public class Event {
 
-	private int eventId;
+	private Long eventId;
 	private String title;
 	private String location;
 	private String startTimeStamp;
 	private String endTimeStamp;
 
-	public int getEventId() {
+	public Long getEventId() {
 		return eventId;
 	}
 
-	public void setEventId(int eventId) {
+	public void setEventId(Long eventId) {
 		this.eventId = eventId;
 	}
